+++ app/CategoryBar.tsx
