+++ app/ProductsPane.js
