+++ app/ct.js
@@ -67,7 +67,7 @@ function _productSearchText(searchText, lang) {
 }
 
 function _productFacetsFilter(facetsValue, productTypeAttributes, lang, targetFacet) {
-  if (!facetsValue) return null;
+  if (!facetsValue || !productTypeAttributes) return null;
   const validFacets = Object.entries(facetsValue)
       .filter(([facetName, _]) => facetName !== targetFacet)
